changed `-K` to `-W`; added same to vdj cellranger script

+++ cellranger_vdj_main.sh
@@ -23,7 +23,8 @@ usage () {
     echo -e "  -F  Path to the centralized fastq dir." 
     echo -e "  -R  Path to the reference dir." 
     echo -e "  -Y  Number of cores for cellranger."    
-    echo -e "  -Z  Amount of memory for cellranger."   
+    echo -e "  -Z  Amount of memory for cellranger."
+    echo -e "  -W  Delete .bam* files. Default is true."   
     echo -e "  -h  This message."
 }
 
@@ -32,6 +33,7 @@ RECEPTOR_SET=false
 PATH_ROOT_SET=false
 PATH_FASTQ_SET=false
 PATH_REF_SET=false
+BOOL_DEL_BAM_SET=false
 
 # Get commandline arguments
 while getopts "J:K:T:F:R:Y:Z:h" OPT; do
@@ -54,6 +56,9 @@ while getopts "J:K:T:F:R:Y:Z:h" OPT; do
     Y)  CR_N=$OPTARG
         ;;
     Z)  CR_M=$OPTARG
+        ;;
+    W)  BOOL_DEL_BAM=$OPTARG
+        BOOL_DEL_BAM_SET=true
         ;;
     h)  usage
         exit
@@ -97,6 +102,10 @@ if ! $PATH_REF_SET; then
     exit 1
 fi
 
+# Set BOOL_DEL_BAM to true if no -W specified
+if ! $BOOL_DEL_BAM_SET; then
+    BOOL_DEL_BAM=true
+fi
 
 # paths
 
@@ -161,7 +170,11 @@ for ((IDX=1; IDX<=${N_LINES}; IDX++)); do
 		&> "${PATH_LOG_ID}"
 
     # remove .bam, .bambi, etc.
-    rm "${PATH_OUTPUT}${CUR_ID}"/outs/*.bam*
+
+    if $BOOL_DEL_BAM; then
+        rm "${PATH_OUTPUT}${CUR_ID}"/outs/*.bam*
+    fi
+
     rm "${PATH_OUTPUT}${CUR_ID}"/_*
     rm -r "${PATH_OUTPUT}${CUR_ID}/SC_VDJ_ASSEMBLER_CS"
     rm "${PATH_OUTPUT}${CUR_ID}/${CUR_ID}.mri.tgz"
